Change domain.py from a module to a package
