+++ clarity_ext/domain/validation.py
@@ -1,0 +1,18 @@
+class ValidationType:
+    ERROR = 1
+    WARNING = 2
+
+
+class ValidationException:
+    def __init__(self, msg, validation_type=ValidationType.ERROR):
+        self.msg = msg
+        self.type = validation_type
+
+    def _repr_type(self):
+        if self.type == ValidationType.ERROR:
+            return "Error"
+        elif self.type == ValidationType.WARNING:
+            return "Warning"
+
+    def __repr__(self):
+        return "{}: {}".format(self._repr_type(), self.msg)
